データ種痘を書き出し
